Use new queue objects + constructor
I though that unshift/shift on big arrays might be expensive as it renumbers
every item in the array, so I made this special queue object on top of an Object object
and inplemented the arrayish methods on it but in a such a way that adding at either
side (unshift/push) is cheap. It does not make a big difference probably because the
queue is never so large that this can make a difference.