Fixing a bug in url.parse: https://github.com/joyent/node/issues/562
The file:// protocol *always* has a hostname; it's frequently abbreviated as
an empty string, which represents 'localhost' implicitly.

According to RFC 1738 (http://tools.ietf.org/html/rfc1738):

A file URL takes the form:

   file://<host>/<path>

where <host> is the fully qualified domain name of the system on
which the <path> is accessible...

As a special case, <host> can be the string "localhost" or the empty
string; this is interpreted as `the machine from which the URL is
being interpreted'.