Revert "removing copy-back-to-sandbox in vm"
After a long discussion with @isaacs, we’ve decided to defer consideration of
API changes to ~0.7, and reduce this change to a relatively safe subset of
changes that don’t affect the API for immediate inclusion in 0.6.

The changes reverted in this commit will be coming back in 0.7, after a
discussion of the alternatives.

This reverts commit 2c1f5206bbb23bd34624acbcfb475dddcda0b883.