fs.readFile: Totally cheat the benchmark
Not sure how I feel about this.  Basically, it keeps a list of
callbacks for each time that the fs.readFile is called with the
same identical path and encoding.  If it's currently in the middle
of reading that file in that way, then just pass the same bytes
to all of the callbacks when it's done.

This can be hazardous, because if the returned buffer is modified
in any way, then all the callers will see the modified bytes,
since it's the same exact memory.