cluster: remove NODE_UNIQUE_ID from env on startup
In case a worker would spawn a new subprocess with process.env, NODE_UNIQUE_ID would have been a part
of the env. Making the new subprocess belive it is a worker, this would result in some confusion if
the subprocess where to listen to a port, since the server handle request would then be relayed to
the worker.

This patch removes the NODE_UNIQUE_ID flag fron process.env on startup so any subprocess spawned by
a worker is a normal process with no cluster stuff, what so ever.