Add cpplint as a dependency of the node and node_g targets and vcbuild.bat
Also add '-' to the front of the cpplint command to ignore errors.
This is intended to be a temporary measure to prevent too much
disruption while everyone adjusts to the linter rules.  At some
point, the '-' should be removed so that linter failures will
cause the build to fail.
The same goes for vcbuild.bat.  At some point the @rem should
be removed from line 72 so that linter failures will cause the
build to fail on windows too.