windows: reduce redundant strlen/wcslen calls
Added two new ETW writer macros that accept an existing string length so
as to not make a redundant strlen or wcslen call.

I'm using the new version of the macro in the one place that I could find
where the string length is already known in context. Other reductions are
probably possible here, but would take more plumbing (and changes in
dtrace code as well).