piped stream should not throw if dest err handler
A piped stream should not throw an error if there is an `error`
listener on the pipe destination, so that error handling is
propagated through all of the pipes to the end where it can
be handled in one place.

```javascript
rstream
  .pipe(foo)
  .pipe(bar)
  .on('error', handleErr);
```

If a source error handler or destination error handler exists
then we should not throw an error, but rather forward the
error to the destination error handler.