Use "Node.js" notation instead of "node.js"
According to https://github.com/joyent/node/wiki/FAQ,
"Node.js" is the correct capitalization.

It would be better to follow the guideline in the title
of html documents.

Other than this "node.js" notation is used in some places.
As there may be room for discussion to change all of them,
this patch has left them untouched.