expose the process global through require('process');
This was done for the node specific buffer module a long time ago,
however process was left as a global. This allows the use of process via
a require so that going forward this can become the practice for
obtaining the process info much like buffer global was phased out.