Streams2 can contain objects
We detect for non-string and non-buffer values in onread and
turn the stream into an "isObjectStream" stream.

If we are in "isObjectStream" mode then howMuchToRead will
always return 1, state.length will always have 1 appended
to it when there is a new item and fromList always takes
the first value from the list.

This means that for object streams, the n in read(n) is
ignored and read() will always return a single value