falsey values in stream (as object stream)
Currently falsey chunks are ignored in object stream. This
has been changed to undefined and null signal eof and are
ignored.

There's also a fix for it to handle "" which was previously
treated as a string / buffer for state.length decremental.

There also tests for correct low / high watermark handling using
.push(...)