timers: consistent this keyword in setImmediate
When calling setImmediate with extra arguments the this keyword in the callback
would refer to the global object, but when not calling setImmediate with extra
arguments this would refer to the returned handle object.

This commit fixes that inconsistency so its always set handle object. The
handle object was chosen for performance reasons.