Streams2-ify `dgram.Socket`.
This is a proof-of-concept implementation of a stream-based `dgram.Socket`.
The class based on `stream.Duplex` with `{objectMode: true}`.

Received UDP packets can be retrieved by calling `read()` and look like
this:

    {
      data: udpDataPayload,
      ip: { src: '1.2.3.4' },
      udp { srcPort: 52 }
    }

UDP packets can be sent by pass object that look like the following to
`write()`:

    {
      data: udpDataPayload,
      ip: { dst: '4.3.2.1' },
      udp: { dstPort: 123 }
    }

Backward compatibility is maintained.  The existing `send()` call can be
used freely.  If `on('message')` is called, then packets will automatically
be drained from the read queue and emitted using the existing `'message'`
event API.

See the new stream-based ping-pong test in `test-dgram-pingpong-stream.js`.