doc: Update doc for http.IncomingMessage.socket
For HTTPS servers, the socket property of the IncomingMessage is a
tls.CleartextStream rather than a net.Socket.  Update the documentation
to reflect this.

Also update the documentation for getting the client authentication
details.  verifyPeer was removed in df46c8e6 2.5 years ago and
request.connection is a now-undocumented alias for request.socket.
Update these to the current names.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>