Enforce system libtool on OSX.
The libtool provided by OSX is special, and incompatible with GNU
libtool.  This primarily manifests itself with

  libtool: unrecognized option `-static'

errors, when the 'libtool' as picked by $PATH is a GNU one.

By hardcoding the path to the system libtool, we can avoid such
errors in both the node build itself as well as in module builds.